chore: move UpdateUserRequest interface to js-client-library. fix typo in Users.test.tsx
