+++ packages/javascript/js-client-library/src/client.ts
@@ -717,19 +717,7 @@ class BHEAPIClient {
             options
         );
 
-    updateUser = (
-        userId: string,
-        user: {
-            firstName: string;
-            lastName: string;
-            emailAddress: string;
-            principal: string;
-            roles: number[];
-            SAMLProviderId?: string;
-            is_disabled?: boolean;
-        },
-        options?: types.RequestOptions
-    ) =>
+    updateUser = (userId: string, user: types.UpdateUserRequest, options?: types.RequestOptions) =>
         this.baseClient.patch(
             `/api/v2/bloodhound-users/${userId}`,
             {
