+++ packages/javascript/js-client-library/src/types.ts
@@ -260,3 +260,13 @@ export interface ClearDatabaseRequest {
     deleteDataQualityHistory: boolean;
     deleteAssetGroupSelectors: number[];
 }
+
+export interface UpdateUserRequest {
+    firstName: string;
+    lastName: string;
+    emailAddress: string;
+    principal: string;
+    roles: number[];
+    SAMLProviderId?: string;
+    is_disabled?: boolean;
+}
